events/2025/pyproject.toml
[project]
name = "2025"
version = "1.0.0"
description = "My solutions to Everybody Codes 2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pycryptodome>=3.23.0",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
]

[tool.ruff.format]
quote-style = "single"
